reinit db

+++ tasks/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from .models import Task
+from django.contrib.auth.decorators import login_required
 from .forms import TaskForm
 
 # Create your views here.
@@ -7,21 +8,23 @@ def index(request):
     return render(request, 'tasks/index.html')
 
 
+@login_required
 def tasks(request):
-    tasks = Task.objects.order_by('date_added')
+    tasks = Task.objects.filter(owner=request.user).order_by("priority",'date_added')
     context = {'tasks': tasks}
     return render(request, 'tasks/tasks.html', context)
 
 
+@login_required
 def task(request, task_id):
     task = Task.objects.get(id=task_id)
-    # Проверка того, что тема принадлежит текущему пользователю.
-    # if topic.owner != request.user:
-    #     raise Http404
+    if task.owner != request.user:
+        raise Http404
     context = {'task': task}
     return render(request, 'tasks/task.html', context)
 
 
+@login_required
 def new_task(request):
     if request.method != 'POST':
         # Данные не отправлялись; создается пустая форма.
@@ -30,15 +33,20 @@ def new_task(request):
         # Отправлены данные POST; обработать данные.
         form = TaskForm(data=request.POST)
         if form.is_valid():
-            form.save()
+            new_topic = form.save(commit=False)
+            new_topic.owner = request.user
+            new_topic.save()
             return redirect('tasks:tasks')
     # Вывести пустую или недействительную форму.
     context = {'form': form}
     return render(request, 'tasks/new_task.html', context)
 
 
+@login_required
 def edit_task(request, task_id):
     task = Task.objects.get(id=task_id)
+    if task.owner != request.user:
+        raise Http404
     if request.method != 'POST':
         # Исходный запрос; форма заполняется данными текущей записи.
         form = TaskForm(instance=task)
@@ -52,8 +60,11 @@ def edit_task(request, task_id):
     return render(request, 'tasks/edit_task.html', context)
 
 
+@login_required
 def delete_task(request, task_id):
     task = Task.objects.get(id=task_id)
+    if task.owner != request.user:
+        raise Http404
     if task:
         task.delete()
     return redirect("tasks:tasks")
